+++ internal/lazy/coordiantion.go
@@ -8,38 +8,46 @@ import (
 	builder "github.com/ydb-platform/ydb-go-sdk/v3/internal/coordination"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/coordination/config"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/database"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
 	"github.com/ydb-platform/ydb-go-sdk/v3/retry"
 	"github.com/ydb-platform/ydb-go-sdk/v3/scheme"
 )
 
 type lazyCoordination struct {
-	db      database.Connection
-	options []config.Option
-	c       coordination.Client
-	m       sync.Mutex
+	db     database.Connection
+	config config.Config
+	c      coordination.Client
+	m      sync.Mutex
 }
 
 func Coordination(db database.Connection, options []config.Option) coordination.Client {
 	return &lazyCoordination{
-		db:      db,
-		options: options,
+		db:     db,
+		config: config.New(options...),
 	}
 }
 
 func (c *lazyCoordination) CreateNode(ctx context.Context, path string, config coordination.NodeConfig) (err error) {
+	if !c.config.AutoRetry() {
+		return c.client().CreateNode(ctx, path, config)
+	}
 	return retry.Retry(ctx, func(ctx context.Context) (err error) {
 		return c.client().CreateNode(ctx, path, config)
 	})
 }
 
 func (c *lazyCoordination) AlterNode(ctx context.Context, path string, config coordination.NodeConfig) (err error) {
+	if !c.config.AutoRetry() {
+		return c.client().AlterNode(ctx, path, config)
+	}
 	return retry.Retry(ctx, func(ctx context.Context) (err error) {
 		return c.client().AlterNode(ctx, path, config)
 	})
 }
 
 func (c *lazyCoordination) DropNode(ctx context.Context, path string) (err error) {
+	if !c.config.AutoRetry() {
+		return c.client().DropNode(ctx, path)
+	}
 	return retry.Retry(ctx, func(ctx context.Context) (err error) {
 		return c.client().DropNode(ctx, path)
 	})
@@ -53,11 +61,14 @@ func (c *lazyCoordination) DescribeNode(
 	config *coordination.NodeConfig,
 	err error,
 ) {
+	if !c.config.AutoRetry() {
+		return c.client().DescribeNode(ctx, path)
+	}
 	err = retry.Retry(ctx, func(ctx context.Context) (err error) {
 		entry, config, err = c.client().DescribeNode(ctx, path)
-		return xerrors.WithStackTrace(err)
+		return err
 	})
-	return entry, config, xerrors.WithStackTrace(err)
+	return entry, config, err
 }
 
 func (c *lazyCoordination) Close(ctx context.Context) (err error) {
@@ -66,21 +77,14 @@ func (c *lazyCoordination) Close(ctx context.Context) (err error) {
 	if c.c == nil {
 		return nil
 	}
-	defer func() {
-		c.c = nil
-	}()
-	err = c.c.Close(ctx)
-	if err != nil {
-		return xerrors.WithStackTrace(err)
-	}
-	return nil
+	return c.c.Close(ctx)
 }
 
 func (c *lazyCoordination) client() coordination.Client {
 	c.m.Lock()
 	defer c.m.Unlock()
 	if c.c == nil {
-		c.c = builder.New(c.db, c.options)
+		c.c = builder.New(c.db, c.config)
 	}
 	return c.c
 }