+++ internal/lazy/scheme.go
@@ -7,22 +7,21 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/database"
 	builder "github.com/ydb-platform/ydb-go-sdk/v3/internal/scheme"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/scheme/config"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
 	"github.com/ydb-platform/ydb-go-sdk/v3/retry"
 	"github.com/ydb-platform/ydb-go-sdk/v3/scheme"
 )
 
 type lazyScheme struct {
-	db      database.Connection
-	options []config.Option
-	c       scheme.Client
-	m       sync.Mutex
+	db     database.Connection
+	config config.Config
+	c      scheme.Client
+	m      sync.Mutex
 }
 
 func Scheme(db database.Connection, options []config.Option) scheme.Client {
 	return &lazyScheme{
-		db:      db,
-		options: options,
+		db:     db,
+		config: config.New(options...),
 	}
 }
 
@@ -42,22 +41,15 @@ func (s *lazyScheme) Close(ctx context.Context) (err error) {
 	if s.c == nil {
 		return nil
 	}
-	defer func() {
-		s.c = nil
-	}()
-	err = s.c.Close(ctx)
-	if err != nil {
-		return xerrors.WithStackTrace(err)
-	}
-	return nil
+	return s.c.Close(ctx)
 }
 
 func (s *lazyScheme) DescribePath(ctx context.Context, path string) (e scheme.Entry, err error) {
 	err = retry.Retry(ctx, func(ctx context.Context) (err error) {
 		e, err = s.client().DescribePath(ctx, path)
-		return xerrors.WithStackTrace(err)
+		return err
 	}, retry.WithIdempotent(true))
-	return e, xerrors.WithStackTrace(err)
+	return e, err
 }
 
 func (s *lazyScheme) MakeDirectory(ctx context.Context, path string) (err error) {
@@ -69,9 +61,9 @@ func (s *lazyScheme) MakeDirectory(ctx context.Context, path string) (err error)
 func (s *lazyScheme) ListDirectory(ctx context.Context, path string) (d scheme.Directory, err error) {
 	err = retry.Retry(ctx, func(ctx context.Context) (err error) {
 		d, err = s.client().ListDirectory(ctx, path)
-		return xerrors.WithStackTrace(err)
+		return err
 	}, retry.WithIdempotent(true))
-	return d, xerrors.WithStackTrace(err)
+	return d, err
 }
 
 func (s *lazyScheme) RemoveDirectory(ctx context.Context, path string) (err error) {
@@ -84,7 +76,7 @@ func (s *lazyScheme) client() scheme.Client {
 	s.m.Lock()
 	defer s.m.Unlock()
 	if s.c == nil {
-		s.c = builder.New(s.db, s.options)
+		s.c = builder.New(s.db, s.config)
 	}
 	return s.c
 }