+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v3.24.0
 * Fixed re-opening case after close lazy-initialized clients
 * Removed dependency of call context for initializing lazy table client
 * Added `config.AutoRetry()` flag with `true` value by default. `config.AutoRetry()` affects how to errors handle in sub-clients calls.